// packages/openapi-fetch/test/http-methods/schemas/query.d.ts
/**
 * This file was auto-generated by openapi-typescript.
 * Do not make direct changes to the file.
 */

export interface paths {
    "/resources/{id}": {
        parameters: {
            query?: never;
            header?: never;
            path: {
                id: number;
            };
            cookie?: never;
        };
        get?: never;
        put?: never;
        post?: never;
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
        query: {
            parameters: {
                query?: never;
                header?: never;
                path: {
                    id: number;
                };
                cookie?: never;
            };
            requestBody: components["requestBodies"]["QueryPayload"];
            responses: {
                200: components["responses"]["QueryResult"];
                /** @description Not Found */
                404: {
                    headers: {
                        [name: string]: unknown;
                    };
                    content: {
                        "application/json": {
                            error?: string;
                        };
                    };
                };
            };
        };
    };
    "/resources-optional": {
        parameters: {
            query?: never;
            header?: never;
            path: {
                id: number;
            };
            cookie?: never;
        };
        get?: never;
        put?: never;
        post?: never;
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
        query: {
            parameters: {
                query?: never;
                header?: never;
                path: {
                    id: number;
                };
                cookie?: never;
            };
            requestBody?: components["requestBodies"]["QueryPayloadOptional"];
            responses: {
                200: components["responses"]["QueryResult"];
            };
        };
    };
}
export type webhooks = Record<string, never>;
export interface components {
    schemas: {
        QueryPayload: {
            ids: number[];
        };
    };
    responses: {
        /** @description OK */
        QueryResult: {
            headers: {
                [name: string]: unknown;
            };
            content: {
                "application/json": {
                    status?: string;
                    data?: Record<string, never>[];
                };
            };
        };
    };
    parameters: never;
    requestBodies: {
        QueryPayload: {
            content: {
                "application/json": components["schemas"]["QueryPayload"];
            };
        };
        QueryPayloadOptional: {
            content: {
                "application/json": components["schemas"]["QueryPayload"];
            };
        };
    };
    headers: never;
    pathItems: never;
}
export type $defs = Record<string, never>;
export type operations = Record<string, never>;